+++ src/main/java/com/diskterminal/proxy/ClientProxy.java
@@ -1,8 +1,10 @@
 package com.diskterminal.proxy;
 
+import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.fml.common.event.FMLInitializationEvent;
 
 import com.diskterminal.ItemRegistry;
+import com.diskterminal.client.BlockHighlightRenderer;
 
 
 public class ClientProxy extends CommonProxy {
@@ -11,5 +13,8 @@ public class ClientProxy extends CommonProxy {
     public void init(FMLInitializationEvent event) {
         super.init(event);
         ItemRegistry.registerModels();
+
+        // Register block highlight renderer
+        MinecraftForge.EVENT_BUS.register(new BlockHighlightRenderer());
     }
 }